added constant pool of mapping relation and update some details...

+++ src/jvm/class/interfaces/interfaces_reader.go
@@ -6,11 +6,11 @@ type InterfacesStructReader struct {
 	class_file_commons.Reader
 }
 
-func (ir *InterfacesStructReader) ReadInterfacesInfos() (int, []uint16) {
+func (ir *InterfacesStructReader) ReadInterfacesInfos() (uint16, []uint16) {
 	ret := *new([]uint16)
 	size := ir.ReadUint16()
 	for i := 0; i < int(size); i++ {
 		ret = append(ret, ir.ReadUint16())
 	}
-	return int(size), ret
+	return size, ret
 }
